Escenarios

+++ controls.js
@@ -18,7 +18,7 @@ function shootBullet(x,y,dx,dy){
       bullet.scale.setTo(2,2);
       if (direccion == -1){
         bullet.scale.x = -2;
-      }; 
+      };
 
     let x,y;
 
@@ -28,7 +28,7 @@ function shootBullet(x,y,dx,dy){
 
     bullet.body.velocity.x = 1000 * direccion;
     bullet.body.velocity.y = (player.body.velocity.y) * .5;
-    
+
     bulletRate = game.time.now + 300;
     //bullet rate
     }
@@ -42,8 +42,30 @@ function resetBullet (bullet) {
 }
 
 function playerKill(player,bullet){
-  player.kill();
   bullet.kill();
+  player.health-=1;
+
+  if(player.health<0){
+    player.kill();
+    player.deleteEntity();
+    game.vidas -= 1;
+    if(game.vidas>0){
+      playerp = new MainPlayer(game,200,200,10);
+      game.camera.follow(playerp);
+    }
+    txtVidas.setText('Vidas: '+game.vidas);
+  }
+
+  if(game.vidas == 0){
+     playerp.kill();
+     //gTile.kill();
+     txt1=game.add.text(200, 250, 'Perdiste', {font:'80px Arial', fill: '#000'});
+     txt1.fixedToCamera=true;
+     txt2=game.add.text(200, 350, 'Nivel: '+game.nivel, {font:'30px Arial', fill: '#000'});
+     txt2.fixedToCamera=true;
+     txt3=game.add.text(200, 400, 'Puntaje: '+game.puntaje, {font:'30px Arial', fill: '#000'});
+     txt3.fixedToCamera=true;
+  }
 }
 
 function bulletKillGround(bullet,ground){
@@ -57,6 +79,12 @@ function enemyKill(bullet,ene){
   if (ene.health <= 0) {
     ene.kill();
     ene.alive = false;
+    game.puntaje+=100;
+    txtPuntaje.setText('Puntaje: '+game.puntaje);
   }
   console.log(ene.health);
+}
+
+function perderVida(){
+
 }
